Move ipfs related functions to utils/ipfs.py Replace references to constants to config.xxx equivalent Add sys.path workaround for location independent import
