+++ metadata/pulling.py
@@ -1,26 +1,18 @@
-import os
+import os, sys
+
+sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
+
 import pandas as pd
 import requests
 import time
 import json
 import argparse
-import sys
 import base64
 from web3 import Web3
 from web3.exceptions import ContractLogicError
 from web3_multicall import Multicall
-import ipfshttpclient
-import re
-import warnings
 
-ABI_ENDPOINT = "https://api.etherscan.io/api?module=contract&action=getabi&address="
-ENDPOINT = ""
-ATTRIBUTES_FOLDER = "raw_attributes"
-IMPLEMENTATION_SLOT = (
-    "0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc"
-)
-IPFS_GATEWAY = ""
-
+from utils import config, ipfs
 
 """
 Smart contract helper methods
@@ -29,7 +21,7 @@ Smart contract helper methods
 
 def get_contract_abi(address):
     # Get contract ABI
-    abi_url = f"{ABI_ENDPOINT}{address}"
+    abi_url = f"{config.ABI_ENDPOINT}{address}"
     response = requests.get(abi_url)
     try:
         abi = json.loads(response.json()["result"])
@@ -37,7 +29,7 @@ def get_contract_abi(address):
     except Exception as err:
         print(f"Failed to get contract ABI from Etherscan: {err}")
         print(f"Falling back to direct ABI checking")
-        if ENDPOINT != "":
+        if config.ENDPOINT != "":
             # We can check the ABI of non-verified Etherscan contracts
             # if they support ERC165 (which most of them do)
             erc165_abi = [
@@ -56,7 +48,7 @@ def get_contract_abi(address):
                 }
             ]
 
-            w3 = Web3(Web3.HTTPProvider(ENDPOINT))
+            w3 = Web3(Web3.HTTPProvider(config.ENDPOINT))
             contract = w3.eth.contract(Web3.toChecksumAddress(address), abi=erc165_abi)
 
             # Array of contract methods that were verified via ERC165
@@ -122,14 +114,14 @@ def get_contract_abi(address):
 
 def get_contract(address, abi):
     # Connect to web3
-    if ENDPOINT == "":
+    if config.ENDPOINT == "":
         print(
             "\nMust enter a Web3 provider. Open this file and set the ENDPOINT and IPFS_GATEWAY constants. See: https://ipfs.github.io/public-gateway-checker/\n"
         )
         print("Optional: Use -web3_provider as a command line argument")
         sys.exit()
 
-    w3 = Web3(Web3.HTTPProvider(ENDPOINT))
+    w3 = Web3(Web3.HTTPProvider(config.ENDPOINT))
 
     # Check if abi contains the tokenURI function
     contract_functions = [func["name"] for func in abi if "name" in func]
@@ -137,7 +129,9 @@ def get_contract(address, abi):
     if "implementation" in contract_functions:
         # Handle case where the contract is a proxy contract
         # Fetch address for the implementation contract
-        impl_contract = w3.toHex(w3.eth.get_storage_at(address, IMPLEMENTATION_SLOT))
+        impl_contract = w3.toHex(
+            w3.eth.get_storage_at(address, config.IMPLEMENTATION_SLOT)
+        )
 
         # Strip the padded zeros from the implementation contract address
         impl_address = "0x" + impl_contract[-40:]
@@ -177,37 +171,13 @@ def get_contract_function(contract, func_name, abi):
         )
 
 
-def format_ipfs_uri(uri):
-    # Reformat IPFS gateway
-    ipfs_1 = "ipfs://"
-    ipfs_2 = "https://ipfs.io/ipfs/"
-    ipfs_3 = "https://gateway.pinata.cloud/ipfs/"
-    ipfs_hash_identifier = "Qm"
-
-    if IPFS_GATEWAY == "":
-        if uri.startswith(ipfs_1):
-            uri = ipfs_2 + uri[len(ipfs_1) :]
-    else:
-        if uri.startswith(ipfs_1):
-            uri = IPFS_GATEWAY + uri[len(ipfs_1) :]
-        elif uri.startswith(ipfs_2):
-            uri = IPFS_GATEWAY + uri[len(ipfs_2) :]
-        elif uri.startswith(ipfs_3):
-            uri = IPFS_GATEWAY + uri[len(ipfs_3) :]
-        elif "pinata" in uri:
-            starting_index_of_hash = uri.find(ipfs_hash_identifier)
-            uri = IPFS_GATEWAY + uri[starting_index_of_hash:]
-
-    return uri
-
-
 def get_contract_uri(contract, token_id, uri_func, abi):
     # Fetch URI from contract
     uri_contract_func = get_contract_function(contract, uri_func, abi)
 
     try:
         uri = uri_contract_func(token_id).call()
-        uri = format_ipfs_uri(uri)
+        uri = ipfs.format_ipfs_uri(uri)
         return uri
     except ContractLogicError as err:
         raise Exception(err)
@@ -215,7 +185,7 @@ def get_contract_uri(contract, token_id, uri_func, abi):
 
 def get_contract_uri_batch(contract, token_ids, uri_func, abi):
     if len(token_ids) > 0:
-        if ENDPOINT == "":
+        if config.ENDPOINT == "":
             print(
                 "You must enter a Web3 provider. This is currently not a command line option. You must open this file and assign a valid provider to the ENDPOINT and IPFS_GATEWAY constants. See: https://ipfs.github.io/public-gateway-checker/"
             )
@@ -225,11 +195,11 @@ def get_contract_uri_batch(contract, token_ids, uri_func, abi):
             uri_contract_func = get_contract_function(contract, uri_func, abi)
             return uri_contract_func(token_id)
 
-        w3 = Web3(Web3.HTTPProvider(ENDPOINT))
+        w3 = Web3(Web3.HTTPProvider(config.ENDPOINT))
         multicall = Multicall(w3.eth)
         multicall_result = multicall.aggregate(list(map(get_func, token_ids)))
         return {
-            x["inputs"][0]["value"]: format_ipfs_uri(x["results"][0])
+            x["inputs"][0]["value"]: ipfs.format_ipfs_uri(x["results"][0])
             for x in multicall_result.json["results"]
         }
     else:
@@ -299,7 +269,7 @@ def fetch_all_metadata(
 ):
 
     # Create raw attribute folder for collection if it doesnt already exist
-    folder = f"{ATTRIBUTES_FOLDER}/{collection}/"
+    folder = f"{config.ATTRIBUTES_FOLDER}/{collection}/"
     if not os.path.exists(folder):
         os.mkdir(folder)
 
@@ -316,25 +286,30 @@ def fetch_all_metadata(
                 break
             except Exception as err:
                 pass
-        cid = infer_cid_from_uri(uri)
+        cid = ipfs.infer_cid_from_uri(uri)
         if cid is not None:
-            uri_base = IPFS_GATEWAY + cid + "/"
+            uri_base = config.IPFS_GATEWAY + cid + "/"
 
     # First try to get all metadata files from ipfs in bulk
-    if uri_base is not None and uri_base.find("ipfs") != -1:
+    if uri_base is not None and ipfs.is_valid_ipfs_uri(uri_base):
         folder_walk = os.walk(folder, topdown=True, onerror=None, followlinks=False)
         _files = next(folder_walk)[2]
 
         if len(_files) == 0:
-            cid = infer_cid_from_uri(uri_base)
+            cid = ipfs.infer_cid_from_uri(uri_base)
             try:
-                fetch_ipfs_folder(collection_name=collection, cid=cid)
+                ipfs.fetch_ipfs_folder(
+                    collection_name=collection,
+                    cid=cid,
+                    parent_folder=config.ATTRIBUTES_FOLDER,
+                    timeout=60,
+                )
                 folder_walk = os.walk(
                     folder, topdown=True, onerror=None, followlinks=False
                 )
                 _files = next(folder_walk)[2]
                 first_file = _files[0]
-                file_suffix = get_file_suffix(first_file)
+                file_suffix = ipfs.get_file_suffix(first_file)
                 bulk_ipfs_success = True
             except Exception:
                 file_suffix = ".json"
@@ -421,7 +396,7 @@ def fetch_all_metadata(
             # Get the metadata URI
             if uri_base is not None:
                 # Build URI from base URI and URI suffix provided
-                uri_base = format_ipfs_uri(uri_base)
+                uri_base = ipfs.format_ipfs_uri(uri_base)
                 if uri_base.endswith("/"):
                     uri_base = uri_base[:-1]
                 if uri_base.endswith("="):
@@ -489,87 +464,6 @@ def fetch_all_metadata(
                 )
 
     return dictionary_list
-
-
-def fetch_ipfs_folder(collection_name, cid, timeout=120):
-    """
-    Given a collection name, a cid and an optional timeout, this function downloads the entire metadata folder from IPFS.
-
-    :param collection_name The collection name to be used as the folder name
-    :type collection_name: str
-    :param cid: The IPFS CID of folder to download
-    :type cid: str
-    :param timeout: Connection timeout (in seconds) when connecting to the API daemon
-    :type timeout: int | None
-    """
-    infura = "/dns/infura-ipfs.io/tcp/5001/https"
-    ipfs_io = "/dns/ipfs.io/tcp/443/https"
-    ipfs_gateway_io = "/dns/gateway.ipfs.io/tcp/443/https"
-    dweb_link = "/dns/dweb.link/tcp/443/https"
-    pinata = "/dns/gateway.pinata.cloud/tcp/443/https"
-    warnings.filterwarnings(
-        "ignore", category=ipfshttpclient.exceptions.VersionMismatch
-    )
-    gateways = [ipfs_gateway_io, infura, dweb_link, ipfs_io, pinata]
-    print("Attempting to download metadata folder from IPFS...\nPlease wait...")
-
-    for gateway in range(len(gateways)):
-        try:
-            client = ipfshttpclient.connect(addr=gateways[gateway], timeout=timeout)
-            client.get(f"/ipfs/{cid}", target=f"./{ATTRIBUTES_FOLDER}/")
-            print(f"Successfully downloaded metadata folder from IPFS")
-            os.rename(
-                f"./{ATTRIBUTES_FOLDER}/{cid}",
-                f"./{ATTRIBUTES_FOLDER}/{collection_name}",
-            )
-            client.close()
-            break
-        except Exception:
-            if gateway < len(gateways) - 1:
-                print(
-                    "Failed to download metadata folder from IPFS. Trying next gateway..."
-                )
-            else:
-                print(
-                    "Failed to download metadata folder from IPFS.\nFalling back to individual file downloads..."
-                )
-            pass
-
-
-def get_file_suffix(filename, token_id="\\d+"):
-    """
-    Given a filename and an optional token_id, this function returns the file suffix.
-    If the file has no extension, an empty string is returned.
-
-    :param filename
-    :type filename: str
-    :param token_id
-    :type token_id: str | int | None
-    :return: file_suffix
-    :rtype: str
-    """
-    regex = rf"^{token_id}(\.(?P<extension>\w+))?$"
-    matches = re.search(regex, filename)
-    if matches and matches.group("extension"):
-        return matches.group(1)
-    return ""
-
-
-def infer_cid_from_uri(uri):
-    """
-    Given a URI, this function returns the CID.
-    Returns None if the CID is not found.
-
-    :param uri
-    :type uri: str
-    :return: cid
-    :rtype: str | None
-    """
-    cid_pattern = r"Qm[a-zA-Z0-9-_]+"
-    matches = re.search(cid_pattern, uri)
-    if matches:
-        return matches.group(0)
-    return None
 
 
 """
@@ -662,7 +556,7 @@ def pull_metadata(args):
     trait_db = pd.DataFrame.from_records(records)
     trait_db = trait_db.set_index("TOKEN_ID")
     print(trait_db.head())
-    trait_db.to_csv(f"{ATTRIBUTES_FOLDER}/{collection}.csv")
+    trait_db.to_csv(f"{config.ATTRIBUTES_FOLDER}/{collection}.csv")
 
 
 if __name__ == "__main__":
@@ -757,7 +651,7 @@ if __name__ == "__main__":
         "-ipfs_gateway",
         type=str,
         default=None,
-        help=f"IPFS gateway. (default: {IPFS_GATEWAY}).",
+        help=f"IPFS gateway. (default: {config.IPFS_GATEWAY}).",
     )
     ARG_PARSER.add_argument(
         "-sleep",
@@ -774,8 +668,8 @@ if __name__ == "__main__":
     ARGS = ARG_PARSER.parse_args()
 
     if ARGS.ipfs_gateway is not None:
-        IPFS_GATEWAY = ARGS.ipfs_gateway
+        config.IPFS_GATEWAY = ARGS.ipfs_gateway
     if ARGS.web3_provider is not None:
-        ENDPOINT = ARGS.web3_provider
+        config.ENDPOINT = ARGS.web3_provider
 
     pull_metadata(ARGS)
